+++ examples/resources/pingfederate_oauth_client_settings/import.sh
@@ -1,0 +1,2 @@
+# singleton resource with fixed id.
+terraform import pingfederate_oauth_client_settings.settings OauthClientSettings